fix(oca): Spec5 审查修复 - 生命期/线缆格式/ID映射/采样率/mDNS/node_id
gstack-review 发现 7 个真实缺陷(均不在原 38/38 测试覆盖内):

C1 main.cpp:oca_bridge 提到与 oca_server 同作用域,消除 daemon 运行期对
   已释放 bridge 的 UAF(原 shared_ptr 局限在 if 块内)。
C2 connector 序列化按 OCAMicro(OcaLiteMediaSourceConnector)重写:
   connectorID u16(非 u32)、pinMap {u16 key,u8 mode,u16 index}(原多 u32
   ownerONo)、connectorStatus {u16,u8 state,u16 errorCode}、GetPorts mode
   INPUT=1/OUTPUT=2(原 0/1)。真实 Fitcan 控制器可解析 GetSourceConnectors。
C3 mDNS TXT 改 avahi_entry_group_add_service_strlst + AvahiStringList:
   原 varargs 版任一中段字段为空(如 ip_addr_sec 恒空)即 NULL 截断,
   mac_addr/device_id/channels/firmware 永不发布。
H1 connector_id(OCA u16)与 daemon_id(0..63)分离:delete/status/CRUD 一律
   用存储的 daemon_id 调 bridge,修复 off-by-one 误删(原 connector_id&0xFF)。
H2 set_sample_rate 同步更新 Config,GetCurrentRate 读回新值不陈旧,
   emit 的 PropertyChanged 与可读值一致,退出时持久化。
M1 bridge 继承 enable_shared_from_this,observer 注册延后到 start(),
   lambda 捕获 weak_from_this:SessionManager 无注销 API,bridge 析构后
   observer 触发即 no-op,避免二次 UAF。
I1 OcaNetwork.GetIDAdvertised 返回 node_id(经新接口 get_device_id()),
   非 ip_addr(IP 由 GetSystemInterfaces 提供)。

验证:./oca-dev.sh build 绿,oca-test 40/40(新增 2 个 bridge 路径用例)。

+++ daemon/oca/classes/media_transport_network_aes67.hpp
@@ -40,7 +40,7 @@ class OcaMediaTransportNetworkAES67 : public OcaMediaTransportNetwork {
                                        ocp1::Writer& rsp) override;
   ExecResult add_sink_connector_impl(ocp1::Reader& req,
                                      ocp1::Writer& rsp) override;
-  ExecResult delete_connector_impl(uint32_t connector_id,
+  ExecResult delete_connector_impl(uint16_t connector_id,
                                    ocp1::Writer& rsp) override;
 
  private:
